build: [#110] AdStorage Test를 위한 세팅
