+++ FullStackCodingBot/Full StackTests/MockAdStorage.swift
@@ -1,0 +1,1 @@
+import Foundation
