+++ FullStackCodingBot/FullStackCodingBot/Model/DTO/AdsInformation.swift
@@ -1,0 +1,11 @@
+import Foundation
+
+struct AdsInformation: Codable {
+    let ads: [Bool]
+    let lastUpdated: Date
+    let gift: Int?
+    
+    static func empty() -> AdsInformation {
+        return AdsInformation(ads: [], lastUpdated: Date(timeIntervalSince1970: 0), gift: nil)
+    }
+}
